+++ java/lib/src/main/java/com/svix/models/ApplicationOut.java
@@ -24,8 +24,8 @@ public class ApplicationOut {
     @JsonProperty private String id;
     @JsonProperty private Map<String, String> metadata;
     @JsonProperty private String name;
-    @JsonProperty private Long rateLimit;
-    @JsonProperty private Long throttleRate;
+    @JsonProperty private Short rateLimit;
+    @JsonProperty private Short throttleRate;
     @JsonProperty private String uid;
     @JsonProperty private OffsetDateTime updatedAt;
 
@@ -117,7 +117,7 @@ public class ApplicationOut {
     }
 
     @Deprecated
-    public ApplicationOut rateLimit(Long rateLimit) {
+    public ApplicationOut rateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
         return this;
     }
@@ -129,16 +129,16 @@ public class ApplicationOut {
      */
     @javax.annotation.Nullable
     @Deprecated
-    public Long getRateLimit() {
+    public Short getRateLimit() {
         return rateLimit;
     }
 
     @Deprecated
-    public void setRateLimit(Long rateLimit) {
+    public void setRateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
     }
 
-    public ApplicationOut throttleRate(Long throttleRate) {
+    public ApplicationOut throttleRate(Short throttleRate) {
         this.throttleRate = throttleRate;
         return this;
     }
@@ -151,11 +151,11 @@ public class ApplicationOut {
      * @return throttleRate
      */
     @javax.annotation.Nullable
-    public Long getThrottleRate() {
+    public Short getThrottleRate() {
         return throttleRate;
     }
 
-    public void setThrottleRate(Long throttleRate) {
+    public void setThrottleRate(Short throttleRate) {
         this.throttleRate = throttleRate;
     }
