+++ java/lib/src/main/java/com/svix/models/IngestEndpointIn.java
@@ -23,7 +23,7 @@ public class IngestEndpointIn {
     @JsonProperty private String description;
     @JsonProperty private Boolean disabled;
     @JsonProperty private Map<String, String> metadata;
-    @JsonProperty private Long rateLimit;
+    @JsonProperty private Short rateLimit;
     @JsonProperty private String secret;
     @JsonProperty private String uid;
     @JsonProperty private URI url;
@@ -96,7 +96,7 @@ public class IngestEndpointIn {
         this.metadata = metadata;
     }
 
-    public IngestEndpointIn rateLimit(Long rateLimit) {
+    public IngestEndpointIn rateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
         return this;
     }
@@ -107,11 +107,11 @@ public class IngestEndpointIn {
      * @return rateLimit
      */
     @javax.annotation.Nullable
-    public Long getRateLimit() {
+    public Short getRateLimit() {
         return rateLimit;
     }
 
-    public void setRateLimit(Long rateLimit) {
+    public void setRateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
     }
