+++ java/lib/src/main/java/com/svix/models/IngestEndpointUpdate.java
@@ -23,7 +23,7 @@ public class IngestEndpointUpdate {
     @JsonProperty private String description;
     @JsonProperty private Boolean disabled;
     @JsonProperty private Map<String, String> metadata;
-    @JsonProperty private Long rateLimit;
+    @JsonProperty private Short rateLimit;
     @JsonProperty private String uid;
     @JsonProperty private URI url;
 
@@ -95,7 +95,7 @@ public class IngestEndpointUpdate {
         this.metadata = metadata;
     }
 
-    public IngestEndpointUpdate rateLimit(Long rateLimit) {
+    public IngestEndpointUpdate rateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
         return this;
     }
@@ -106,11 +106,11 @@ public class IngestEndpointUpdate {
      * @return rateLimit
      */
     @javax.annotation.Nullable
-    public Long getRateLimit() {
+    public Short getRateLimit() {
         return rateLimit;
     }
 
-    public void setRateLimit(Long rateLimit) {
+    public void setRateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
     }
