+++ java/lib/src/main/java/com/svix/models/OperationalWebhookEndpointIn.java
@@ -26,7 +26,7 @@ public class OperationalWebhookEndpointIn {
     @JsonProperty private Boolean disabled;
     @JsonProperty private Set<String> filterTypes;
     @JsonProperty private Map<String, String> metadata;
-    @JsonProperty private Long rateLimit;
+    @JsonProperty private Short rateLimit;
     @JsonProperty private String secret;
     @JsonProperty private String uid;
     @JsonProperty private URI url;
@@ -127,7 +127,7 @@ public class OperationalWebhookEndpointIn {
         this.metadata = metadata;
     }
 
-    public OperationalWebhookEndpointIn rateLimit(Long rateLimit) {
+    public OperationalWebhookEndpointIn rateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
         return this;
     }
@@ -138,11 +138,11 @@ public class OperationalWebhookEndpointIn {
      * @return rateLimit
      */
     @javax.annotation.Nullable
-    public Long getRateLimit() {
+    public Short getRateLimit() {
         return rateLimit;
     }
 
-    public void setRateLimit(Long rateLimit) {
+    public void setRateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
     }
