+++ java/lib/src/main/java/com/svix/models/OperationalWebhookEndpointOut.java
@@ -29,7 +29,7 @@ public class OperationalWebhookEndpointOut {
     @JsonProperty private Set<String> filterTypes;
     @JsonProperty private String id;
     @JsonProperty private Map<String, String> metadata;
-    @JsonProperty private Long rateLimit;
+    @JsonProperty private Short rateLimit;
     @JsonProperty private String uid;
     @JsonProperty private OffsetDateTime updatedAt;
     @JsonProperty private URI url;
@@ -168,7 +168,7 @@ public class OperationalWebhookEndpointOut {
         this.metadata = metadata;
     }
 
-    public OperationalWebhookEndpointOut rateLimit(Long rateLimit) {
+    public OperationalWebhookEndpointOut rateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
         return this;
     }
@@ -179,11 +179,11 @@ public class OperationalWebhookEndpointOut {
      * @return rateLimit
      */
     @javax.annotation.Nullable
-    public Long getRateLimit() {
+    public Short getRateLimit() {
         return rateLimit;
     }
 
-    public void setRateLimit(Long rateLimit) {
+    public void setRateLimit(Short rateLimit) {
         this.rateLimit = rateLimit;
     }
